swift json-ast: add typed ast and converter
